Move CastlingRights into indexing.rs
Moved CastlingRights into indexing.rs.
