+++ position/src/lib.rs
@@ -5,7 +5,6 @@
 */
 mod bitboard;
 mod board;
-pub mod castling_rights;
 pub mod indexing;
 pub mod move_generation;
 pub mod position;
