+++ position/src/prelude.rs
@@ -1,5 +1,6 @@
-pub use crate::castling_rights::CastlingRights;
-pub use crate::indexing::{Color, File, NegDiag, Piece, PosDiag, Promotion, Rank, Square};
+pub use crate::indexing::{
+    CastlingRights, Color, File, NegDiag, Piece, PosDiag, Promotion, Rank, Square,
+};
 pub use crate::move_generation::{ChessMove, ChessMoveHint, ChessUnmove, LanMove};
 pub use crate::position::{ParseFenError, Position};
 pub use crate::position_hash::PositionHash;
